@@ -1,15 +1,14 @@
 ﻿using ApprovalTests;
-using ApprovalTests.Maintenance;
 using ApprovalTests.Reporters;
 using ApprovalTests.Writers;
-
+using Nemesis.TextParsers.CodeGen.Deconstructable;
 using static Nemesis.TextParsers.CodeGen.Tests.CodeGenUtils;
 
 namespace Nemesis.TextParsers.CodeGen.Tests.ApprovalTests;
 
 [TestFixture, Explicit]
 [UseReporter(typeof(VisualStudioReporter), typeof(ClipboardReporter))]
-internal class AutoDeconstructableGeneratorApprovalTests
+internal class AutoDeconstructable_AT
 {
     [Test] public void ApprovalTestsRecord() => RunCase("Record");
 
@@ -22,25 +21,20 @@ internal class AutoDeconstructableGeneratorApprovalTests
     [Test] public void ApprovalTestsSimpleWrapperStruct() => RunCase("SimpleWrapperStruct");
 
 
-    [Test]
-    public void HouseKeeping() => ApprovalMaintenance.VerifyNoAbandonedFiles();
-
     private static void RunCase(string index)
     {
-        var (_, source, _) = EndToEndCases.GetAutoDeconstructableCases().SingleOrDefault(t => t.name == index);
+        var (_, source, _) = AutoDeconstructableTests.GetAutoDeconstructableCases().SingleOrDefault(t => t.name == index);
         Assert.That(source, Is.Not.Null);
         Assert.That(source, Is.Not.Empty);
 
         var compilation = CreateValidCompilation(source);
 
-        var generatedTrees = GetGeneratedTreesOnly(compilation);
+        var generatedTrees = GetGeneratedTreesOnly(compilation, new AutoDeconstructableGenerator(), AutoDeconstructableGenerator.ATTRIBUTE_NAME);
 
         var actual = ScrubGeneratorComments(generatedTrees.Single());
 
-        actual = Normalize(actual);
+        actual = NormalizeNewLine(actual);
 
         Approvals.Verify(WriterFactory.CreateTextWriter(actual, "cs"));
     }
-
-    private static string Normalize(string text) => text.Replace("\r\n", "\n").Replace("\r", "\n");
 }
